requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
Pillow==9.5.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
